+++ CHANGELOG.md
@@ -31,7 +31,8 @@ Changelog
 * Removed restriction *LinkAddress*, since it is not needed
 * Added support for [MEID](http://en.wikipedia.org/wiki/Mobile_equipment_identifier) and removed restrictions *getNetworkType* and *getPhoneType* ([issue](/../../issues/1843))
 * Prefixed *getAddress* and *getBondedDevices* with *Bluetooth*
-* Prefixed *setPreviewCallback*, *setPreviewCallbackWithBuffer*, *setOneShotPreviewCallback* and *takePicture* with *Camera*
+* Replaced *setPreviewCallback*, *setPreviewCallbackWithBuffer*, *setOneShotPreviewCallback* by *startPreview*
+* Prefixed *startPreview* and *takePicture* with *Camera*
 * Prefixed *startRecording* with *Audio*
 * Renamed *setOutputFile* into *MediaRecorder.start*
 * Added Croatian translation